% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchSort.R
\docType{data}
\name{batchSort}
\alias{batchSort}
\title{A toy example of metabolomic data set sorted according to batch order}
\format{A data frame with 100 rows (metabolites) and 32 variables (samples of patients, controls and QCs sorted according to batch order)}
\usage{
batchSort
}
\description{
A dataset containing the information about 100 metabolites in samples from 14 patients with one specific disease and 10 healthy controls. Quality control samples (QCs) are also included. All samples are sorted according to batch order.
}
\keyword{datasets}
