% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigQC.R
\name{bigQC}
\alias{bigQC}
\title{Quality control samples (QCs) checking}
\usage{
bigQC(data, name, groupnames)
}
\arguments{
\item{data}{Data table with variables (metabolites) in columns. Samples in rows are sorted according to specific groups.}

\item{name}{A character string or expression indicating a name of data set. It occurs in names of every output.}

\item{groupnames}{A character vector defining specific groups in data. Every string must be specific for each group and they must not overlap.}
}
\value{
Boxplots of QCs and the other data groups.

Excel file with the list of questionable variables from two steps of evaluation.
}
\description{
Quality control samples (QCs) are checked to data irregularities. It is used for data from untargeted metabolomic analysis.
}
\details{
Values of QCs are evaluated and questionable values for particular variables are denoted. There are two steps of evaluation: 1. QCs with completely higher values than the maximum of data, 2. QCs higher than majority of data.

Up to twenty different groups can be distinguished in data (including QCs).
}
\examples{
data=metabol
name="Metabolomics"    #name of the project
groupnames=c("Con","Pat","QC")
bigQC(data,name,groupnames)
}
