% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphsBoxPvalue.R
\name{GraphsBoxPvalue}
\alias{GraphsBoxPvalue}
\title{Boxplots with p-values}
\usage{
GraphsBoxPvalue(data, name, groupnames, type = "par", tsf = "clr",
  QCs = FALSE)
}
\arguments{
\item{data}{Data table with variables (metabolites) in columns. Samples in rows are sorted according to specific groups.}

\item{name}{A character string or expression indicating a name of data set. It occurs in names of every ouput.}

\item{groupnames}{A character vector defining specific groups in data. Every string must be specific for each group and they must not overlap.}

\item{type}{Parametric ("par" - default) or nonparametric ("nonpar") test must be chosen.}

\item{tsf}{Data transformation must be defined by "clr" (default), "log", "log10", "PQN", "lnPQN", "pareto" or "none". See Details.}

\item{QCs}{logical. If FALSE (default) quality control samples (QCs) are automatically distinguished and skipped.}
}
\value{
Boxplot graphs with p-values and stars representing p-values.

Excel file with medians and differences of these medians of all groups in data.
}
\description{
Display 4 variants of boxplots with p-values from t-test or Wilcoxon test in PDF files for every variable (metabolite) in data table separately.
}
\details{
Data transformation: with "clr" clr trasformation is used (see References), with "log" natural logarithm is used, with "log10" decadic logarithm is used, with "pareto" data are only scaled by Pareto scaling, with "PQN" probabilistic quotient normalization is done, with "lnPQN" natural logarithm of PQN transformed data is done, with "none" no tranformation is done.

Up to nine different groups can be distinguished in data.
}
\examples{
data=metabol
name="Metabolomics"    #name of the project
groupnames=c("Con","Pat","QC")
GraphsBoxPvalue(data,name,groupnames)
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). p. 416.
}
