% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphsCorr.R
\name{GraphsCorr}
\alias{GraphsCorr}
\title{Correlation graphs}
\usage{
GraphsCorr(data, name, groupnames, tsf = "clr", QCs = FALSE)
}
\arguments{
\item{data}{Data table with variables (metabolites) in columns. Samples in rows are sorted according to specific groups.}

\item{name}{A character string or expression indicating a name of data set. It occurs in names of every output.}

\item{groupnames}{A character vector defining specific groups in data. Every string must be specific for each group and they must not overlap.}

\item{tsf}{Data transformation must be defined by "clr" (default), "log", "log10", "PQN", "lnPQN", "pareto" or "none". See Details.}

\item{QCs}{logical. If FALSE (default) quality control samples (QCs) are automatically distinguished and skipped.}
}
\value{
Correlation graph.

Excel file with Pearson and Spearman correlations in standard tables and also sorted in one column. Significance tests are also evaluated.
}
\description{
Evaluate correlations (Pearson and Spearman) between all pairs of variables, display correlation plots.
}
\details{
Data transformation: with "clr" clr trasformation is used (see References), with "log" natural logarithm is used, with "log10" decadic logarithm is used, with "pareto" data are only scaled by Pareto scaling, with "PQN" probabilistic quotient normalization is done, with "lnPQN" natural logarithm of PQN transformed data is done, with "none" no tranformation is done.

Up to twenty different groups can be distinguished in data (including QCs).

Cor_S denotes Spearman correlation coefficient.
}
\examples{
data=metabol
name="Metabolomics"    #name of the project
groupnames=c("Con","Pat","QC")
GraphsCorr(data,name,groupnames)
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). p. 416.
}
