% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphsPermut.R
\name{GraphsPermut}
\alias{GraphsPermut}
\title{Permutation test of PLS-DA and OPLS-DA}
\usage{
GraphsPermut(data, name, groupnames, tsf = "clr", type = "opls",
  permut = 100, QCs = FALSE)
}
\arguments{
\item{data}{Data table with variables (metabolites) in columns. Samples in rows are sorted according to specific groups.}

\item{name}{A character string or expression indicating a name of data set. It occurs in names of every output.}

\item{groupnames}{A character vector defining specific groups in data. Every string must be specific for each group and they must not overlap.}

\item{tsf}{Data transformation must be defined by "clr" (default), "log", "log10", "PQN", "lnPQN", "pareto" or "none". See Details.}

\item{type}{Definition which type of method do you want to use. If "opls" is set (default), OPLS-DA is performed and the number of orthogonal components is automatically computed by using the cross-validation (with a maximum of 9 orthogonal components). If two groups are not nicely separated, the algorithm returns error. In this case set this parameter to "err".  When "pls" is set, PLS-DA is done.}

\item{permut}{Number of random permutations of response labels to estimate R2Y and Q2Y significance by permutation testing. Default is 100.}

\item{QCs}{logical. If FALSE (default) quality control samples (QCs) are automatically distinguished and skipped.}
}
\value{
Score plots and permutation plots of OPLS-DA.

Excel file with summary.
}
\description{
Makes permutation test of PLS-DA and OPLS-DA, displays score plots (for OPLS-DA only), permutation plots and summary table.
}
\details{
Data transformation: with "clr" clr trasformation is used (see References), with "log" natural logarithm is used, with "log10" decadic logarithm is used, with "pareto" data are only scaled by Pareto scaling, with "PQN" probabilistic quotient normalization is done, with "lnPQN" natural logarithm of PQN transformed data is done, with "none" no tranformation is done.

Up to twenty different groups can be distinguished in data (including QCs).
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). p. 416.

Thevenot E. A. et al., Analysis of the human adult urinary metabolome variations with age, body mass index and gender by implementing a comprehensive workflow for univariate and OPLS statistical analyses,Journal of Proteome Research 14, 3322-3335, 2015.
}
