% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphsPLSDA.R
\name{GraphsPLSDA}
\alias{GraphsPLSDA}
\title{Partial least squares - discriminant analysis (PLS-DA)}
\usage{
GraphsPLSDA(data, name, groupnames, type = "points", tsf = "clr",
  top = 20, QCs = TRUE)
}
\arguments{
\item{data}{Data table with variables (metabolites) in columns. Samples in rows are sorted according to specific groups.}

\item{name}{A character string or expression indicating a name of data set. It occurs in names of every output.}

\item{groupnames}{A character vector defining specific groups in data. Every string must be specific for each group and they must not overlap.}

\item{type}{A type of plots must be defined by "points" (default), "names" or "both".}

\item{tsf}{Data transformation must be defined by "clr" (default), "log", "log10", "PQN", "lnPQN", "pareto" or "none". See Details.}

\item{top}{How many rays with highest lengths should be in biplot? The default is 20.}

\item{QCs}{logical. If TRUE (default) quality control samples (QCs) are automatically distinguished. See Details.}
}
\value{
Score plot and biplot of PLS-DA. Cross validation plot.

Excel file with lengths of rays in biplot and mean square error of prediction (MSEP) of data.
}
\description{
Makes partial least squares - discriminant analysis (PLS-DA), displays score plots, loading plots and biplots.
}
\details{
Data transformation: with "clr" clr trasformation is used (see References), with "log" natural logarithm is used, with "log10" decadic logarithm is used, with "pareto" data are only scaled by Pareto scaling, with "PQN" probabilistic quotient normalization is done, with "lnPQN" natural logarithm of PQN transformed data is done, with "none" no tranformation is done.

Up to twenty different groups can be distinguished in data (including QCs).

If quality control samples (QCs) are present in data and QCs=TRUE, versions with QCs and without them are displayed. If QCs=TRUE and QCs are not present in data, this step is automatically skipped.
}
\examples{
data=metabol
name="Metabolomics"    #name of the project
groupnames=c("Con","Pat","QC")
GraphsPLSDA(data,name,groupnames)
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). p. 416.
}
