% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphsROCcomb.R
\name{GraphsROCcomb}
\alias{GraphsROCcomb}
\title{ROC curves for combination of variables}
\usage{
GraphsROCcomb(data, name, groupnames, selm, selg = c(1, 2),
  tsf = "clr", QCs = FALSE)
}
\arguments{
\item{data}{Data table with variables (metabolites) in columns. Samples in rows are sorted according to specific groups.}

\item{name}{A character string or expression indicating a name of data set. It occurs in names of every output.}

\item{groupnames}{A character vector defining specific groups in data. Every string must be specific for each group and they must not overlap.}

\item{selm}{Select sequence number of metabolites, their combination by logistic regression will be plotted in ROC curve. Up to six metabolites can be chosen.}

\item{selg}{Select sequence number of groups of data for which ROC curve will be plotted. The dafult is c(1,2).}

\item{tsf}{Data transformation must be defined by "clr" (default), "log", "log10", "PQN", "lnPQN", "pareto" or "none". See Details.}

\item{QCs}{logical. If FALSE (default) quality control samples (QCs) are automatically distinguished and skipped.}
}
\value{
ROC curve of chosen combination of variables with confidence intervals of area under the curve (AUC).
}
\description{
Display ROC curves with confidence intervals for chosen combination of variables. Variables are combined by logistic regression.
}
\details{
Data transformation: with "clr" clr trasformation is used (see References), with "log" natural logarithm is used, with "log10" decadic logarithm is used, with "pareto" data are only scaled by Pareto scaling, with "PQN" probabilistic quotient normalization is done, with "lnPQN" natural logarithm of PQN transformed data is done, with "none" no tranformation is done.
}
\examples{
data=metabol
name="Metabolomics"    #name of the project
groupnames=c("Con","Pat","QC")
GraphsROCcomb(data,name,groupnames,selm=c(2,3))
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). p. 416.
}
