% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphsVIPOPLSDA.R
\name{GraphsVIPOPLSDA}
\alias{GraphsVIPOPLSDA}
\title{Orthogonal partial least squares - discriminant analysis (OPLS-DA)}
\usage{
GraphsVIPOPLSDA(data, name, groupnames, tsf = "clr", top = 30,
  qu = 0.75, QCs = FALSE)
}
\arguments{
\item{data}{Data table with variables (metabolites) in columns. Samples in rows are sorted according to specific groups.}

\item{name}{A character string or expression indicating a name of data set. It occurs in names of every output.}

\item{groupnames}{A character vector defining specific groups in data. Every string must be specific for each group and they must not overlap.}

\item{tsf}{Data transformation must be defined by "clr" (default), "log", "log10", "PQN", "lnPQN", "pareto" or "none". See Details.}

\item{top}{How many most important variables (in absolute values)  should be highlighted in s-plot? The default is 30.}

\item{qu}{Which quantile of the important variables (in absolute values) should be highlighted in s-plot? The default is 0.75.}

\item{QCs}{logical. If FALSE (default) quality control samples (QCs) are automatically distinguished and skipped.}
}
\value{
Score plot and s-plots of OPLS-DA.

Excel file with s-plot summary for every variable.
}
\description{
Makes orthogonal partial least squares - discriminant analysis (OPLS-DA), displays score plots and s-plots.
}
\details{
Data transformation: with "clr" clr trasformation is used (see References), with "log" natural logarithm is used, with "log10" decadic logarithm is used, with "pareto" data are only scaled by Pareto scaling, with "PQN" probabilistic quotient normalization is done, with "lnPQN" natural logarithm of PQN transformed data is done, with "none" no tranformation is done.

S-plots can be used only for comparison of two groups. If there is more groups in data, all possible combinations of pairs are evaluated.

Up to twenty different groups can be distinguished in data (including QCs).
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). p. 416.

Gaude, E.et al. (2012) muma: Metabolomics Univariate and Multivariate Analysis. R package version 1.4. \url{https://CRAN.R-project.org/package=muma}
}
