% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphsVolcano.R
\name{GraphsVolcano}
\alias{GraphsVolcano}
\title{Logratio volcano plots}
\usage{
GraphsVolcano(data, name, groupnames, type = "par", tsf = "clr",
  pair = FALSE, QCs = FALSE, hy = (0.05/length(colnames(data))),
  hx = FALSE, hhx = 1.5)
}
\arguments{
\item{data}{Data table with variables (metabolites) in columns. Samples in rows are sorted according to specific groups.}

\item{name}{A character string or expression indicating a name of data set. It occurs in names of every output.}

\item{groupnames}{A character vector defining specific groups in data. Every string must be specific for each group and they must not overlap.}

\item{type}{Parametric ("par" - default) or nonparametric ("nonpar") test must be chosen.}

\item{tsf}{Data transformation must be defined by "clr" (default), "log", "log10", "PQN", "lnPQN", "pareto" or "none". See Details.}

\item{pair}{logical. If TRUE then the paired tests are done. Default is FALSE.}

\item{QCs}{logical. If FALSE (default) quality control samples (QCs) are automatically distinguished and skipped.}

\item{hy}{Important points are evaluted on y-axis with this limit. The default is alpha=0.05 with Bonferroni correction.}

\item{hx}{logical. If TRUE important points are evaluated from both, x-axis and y-axis. Default is FALSE.}

\item{hhx}{If hx=TRUE, set the absolute value of the limit for most important differences of medians (deafult is 1.5).}
}
\value{
Volcano plots.

Excel sheet with differences of medians and p-values from chosen test.
}
\description{
Display logratio volcano plot comparing two specific groups.
}
\details{
Data transformation: with "clr" clr trasformation is used (see References), with "log" natural logarithm is used, with "log10" decadic logarithm is used, with "pareto" data are only scaled by Pareto scaling, with "PQN" probabilistic quotient normalization is done, with "lnPQN" natural logarithm of PQN transformed data is done, with "none" no tranformation is done.

Logratio volcano plot can be used only for comparison of two groups. If there is more groups in data, all possible combinations of pairs are evaluated.

Logratio volcano plot is a version of standard volcano plot with difference of group medians on x-axis (only for tsf=c("clr","log")) and -log10 of p-value from t-test or Wilcoxon test on y-axis. For "pareto" transformation standard volcano plot is done with log2(fold change) on x-axis.
}
\examples{
data=metabol
name="Metabolomics"    #name of the project
groupnames=c("Con","Pat","QC")
GraphsVolcano(data,name,groupnames)
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). p. 416.
}
