% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabol.R
\docType{data}
\name{metabol}
\alias{metabol}
\title{A toy example of metabolomic data set}
\format{A data frame with 13 rows (samples) and 50 variables (metabolites)}
\usage{
metabol
}
\description{
A small dataset containing the information about 50 metabolites in samples from 5 patients with one specific disease and 5 healthy controls. Quality control samples (QCs) are also included.
}
\keyword{datasets}
