% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preproc.R
\name{Preproc}
\alias{Preproc}
\title{Data preprocessing}
\usage{
Preproc(data, name, groupnames, CVlim = 30)
}
\arguments{
\item{data}{Data table sorted according to batch order (in columns) with variables (metabolites) in rows.}

\item{name}{A character string or expression indicating a name of data set. It occurs in names of every output.}

\item{groupnames}{A character vector defining specific groups in data. Every string must be specific for each group and they must not overlap.}

\item{CVlim}{Limit of coefficient of variation (CV) of quality control samples (QCs). All metabolites with CV of QCs higher than this limit will be omitted from the analysis. The default value is 30.}
}
\value{
LOESS curves.

Excel file with preprocessed data.
}
\description{
Counting zeros, interpolation by LOESS, checking of coefficient of variation (CV) of quality control samples (QCs), zero imputatuion, summary of data preprocessing.
}
\details{
Up to fifteen different groups can be distinguished in data (including QCs).
}
\examples{
data=batchSort
name="Preprocessing"    #name of the project
groupnames=c("Con","Pat","QC")
Preproc(data,name,groupnames)
}
