% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{Tests}
\alias{Tests}
\title{Standard tests}
\usage{
Tests(data, name, groupnames, tsf = "clr", pair = FALSE, QCs = FALSE)
}
\arguments{
\item{data}{Data table with variables (metabolites) in columns. Samples in rows are sorted according to specific groups.}

\item{name}{A character string or expression indicating a name of data set. It occurs in names of every output.}

\item{groupnames}{A character vector defining specific groups in data. Every string must be specific for each group and they must not overlap.}

\item{tsf}{Data transformation must be defined by "clr" (default), "log", "log10", "PQN", "lnPQN", "pareto" or "none". See Details.}

\item{pair}{logical. If TRUE then the paired tests are done. Default is FALSE.}

\item{QCs}{logical. If FALSE (default) quality control samples (QCs) are automatically distinguished and skipped.}
}
\value{
Excel file with p-values from Shapiro-Wilk test, ANOVA, Kruskal-Wallis test, t-test for every variable and Wilcoxon test for every variable and their summary.
}
\description{
Makes Shapiro-Wilk test, ANOVA, Kruskal-Wallis test, t-test for every variable and Wilcoxon test for every variable.
}
\details{
Data transformation: with "clr" clr trasformation is used (see References), with "log" natural logarithm is used, with "log10" decadic logarithm is used, with "pareto" data are only scaled by Pareto scaling, with "PQN" probabilistic quotient normalization is done, with "lnPQN" natural logarithm of PQN transformed data is done, with "none" no tranformation is done.
}
\examples{
data=metabol
name="Metabolomics"    #name of the project
groupnames=c("Con","Pat","QC")
Tests(data,name,groupnames)
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). p. 416.
}
